Move EndSessionPopUp to frontend
