+++ src/main/java/io/github/tkjonesy/utils/models/FileSession.java
@@ -4,7 +4,7 @@ import io.github.tkjonesy.ONNX.models.Log;
 import io.github.tkjonesy.ONNX.models.OnnxRunner;
 import io.github.tkjonesy.frontend.App;
 import io.github.tkjonesy.utils.settings.ProgramSettings;
-import io.github.tkjonesy.utils.EndSessionPopUp;
+import io.github.tkjonesy.frontend.models.EndSessionPopUp;
 import lombok.Getter;
 
 import org.bytedeco.opencv.opencv_core.Mat;
